@AlariCode 2020-04-24
@mjarmoc added your code to master branch and published in nestjs-rmq@1.7.0 (faster then merging your PR, because of code conflicts). Thanks for contributing!